Split Toolshed.HTTP into per-fun mods
